save refresh token and add platform service
